% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{check_tip_spelling}
\alias{check_tip_spelling}
\title{Check spelling of tips}
\usage{
check_tip_spelling(tips, ignore = readLines(system.file("extdata",
  "wordlist", package = "rlangtip")))
}
\arguments{
\item{tips}{data.frame of tips}

\item{ignore}{character vector with words which will be added to the \link[hunspell:dictionary]{hunspell::dictionary}}
}
\description{
Check spelling of tips
}
