% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{close_tips}
\alias{close_tips}
\title{Return a table of close tips}
\usage{
close_tips(tips, cutoff = 0.15, method = "jw", ...)
}
\arguments{
\item{tips}{data.frame of tips}

\item{cutoff}{the cutoff to use}

\item{method}{Method for distance calculation. The default is \code{"osa"},
see \code{\link{stringdist-metrics}}.}

\item{...}{Additional parameters passed to [stringdist::stringdistmatrix]}
}
\description{
Return a table of close tips
}
