% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{filter_dupes}
\alias{filter_dupes}
\title{Filter Dupes}
\usage{
filter_dupes(tbl, cutoff = -3)
}
\arguments{
\item{tbl}{A tibble}

\item{cutoff}{The Z-score cutoff to use for filtering}
}
\description{
Filter Dupes
}
