% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_submissions.R
\name{format_submissions}
\alias{format_submissions}
\title{Add google sheet form data to csv file}
\usage{
format_submissions(subs, tips)
}
\arguments{
\item{subs}{Tibble. Name of tibble containing submission data (output of get_submissions)}

\item{tips}{Tibble. Name of tips table (inst/extdata/tips.csv). Used to identify id ID range.}
}
\description{
Add google sheet form data to csv file
}
