% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipleine.R
\name{get_bunch_o_tweets}
\alias{get_bunch_o_tweets}
\title{Get a bunch of tweets}
\usage{
get_bunch_o_tweets(n_tweets = 5, weighted = TRUE)
}
\arguments{
\item{n_tweets}{Number of tweets to choose}

\item{weighted}{Should tweets with higher scores be chosen more frequently?}
}
\description{
Get a bunch of tweets
}
