% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submissions.R
\name{get_submissions}
\alias{get_submissions}
\title{Import Google Sheets Data}
\usage{
get_submissions(key, tab)
}
\arguments{
\item{key}{Character. The key value associated with the google sheet which has the form responses.}

\item{tab}{Character. Name of the google sheet tab with the form responses.}
}
\description{
Run googlesheets:gs_ls() to initiate authentication
}
