% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets.R
\name{get_tweet_number}
\alias{get_tweet_number}
\title{Get tweet number}
\usage{
get_tweet_number(buffer = 100)
}
\arguments{
\item{buffer}{Numeric. The number to add to the last number of tweets pulled in such that we always ask for more tweets than necessary the next time the script is run.}
}
\description{
Get tweet number
}
