% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets.R
\name{get_tweets}
\alias{get_tweets}
\title{Get Tweets}
\usage{
get_tweets(save_number = TRUE, n_tweets_to_grab = "all")
}
\arguments{
\item{save_number}{Whether to save the numbers of the tweets or not}

\item{n_tweets_to_grab}{Number of tweets to grab, or "all"}
}
\description{
Get Tweets
}
