% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_intake.R
\name{retrieveSubmissions}
\alias{retrieveSubmissions}
\title{Import Google Sheets Data}
\usage{
retrieveSubmissions(key, tab)
}
\arguments{
\item{key}{Character. The key value associated with the google sheet which has the form responses.}

\item{tab}{Character. Name of the google sheet tab with the form responses.}
}
\description{
Import Google Sheets Data
}
