% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtip.R
\name{rtip}
\alias{rtip}
\title{Display an R tip}
\usage{
rtip(id, cowsay = TRUE, color = FALSE)
}
\arguments{
\item{id}{numeric ID of tip to display. If not provided, display a randomly-selected tip.}

\item{cowsay}{if TRUE, a random cowsay animal will present the tip}

\item{color}{if TRUE, use a colorful cowsay display. Ignored if cowsay=FALSE}
}
\description{
Display an R tip
}
