% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{similar_text}
\alias{similar_text}
\title{Return a table of close tips}
\usage{
similar_text(text, cutoff = 0.15, method = "jw", ...)
}
\arguments{
\item{text}{data.frame of tips}

\item{cutoff}{the Z score cutoff to use}

\item{method}{Method for distance calculation. The default is \code{"osa"},
see \code{\link{stringdist-metrics}}.}

\item{...}{Additional parameters passed to [stringdist::stringdistmatrix]}
}
\description{
Return a table of close tips
}
