% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets.R
\name{add_rstats_hashtag}
\alias{add_rstats_hashtag}
\title{Add #rstats}
\usage{
add_rstats_hashtag(tbl, col)
}
\arguments{
\item{tbl}{Tip text.}

\item{col}{Column to add hashtags to.}
}
\description{
Add #rstats
}
\examples{

tibble::tibble(text = "foo") \%>\% add_rstats_hashtag(text)
tibble::tibble(text = "foo #rstats") \%>\% add_rstats_hashtag(text)

}
