% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipleine.R
\name{get_bunch_o_tweets}
\alias{get_bunch_o_tweets}
\title{Get a bunch of tweets}
\usage{
get_bunch_o_tweets(from = "joined", n_tweets = "all",
  weighted = TRUE)
}
\arguments{
\item{from}{Which data source should be used. "canonical" for tips CSV, "twitter" for full Twitter data, "joined" for canonical joined on Twitter data and deduped}

\item{n_tweets}{Number of tweets to choose, or "all"}

\item{weighted}{Should tweets with higher scores be chosen more frequently? Will be false if from is "canonical"}
}
\description{
Get a bunch of tweets
}
